+++ aim/web/ui/src/components/MediaPanel/MediaPanel.d.ts
@@ -9,6 +9,7 @@ import {
 import { IImagesExploreAppConfig } from 'types/services/models/imagesExplore/imagesExploreAppModel';
 
 import { ChartTypeEnum } from 'utils/d3';
+import { SortFields } from 'utils/getSortedFields';
 
 import { MediaTypeEnum } from './config';
 
@@ -33,4 +34,6 @@ export interface IMediaPanelProps {
   tooltipType?: ChartTypeEnum;
   onActivePointChange?: (activePoint: any, focusedStateActive: boolean) => void;
   getBlobsData: (uris: string[]) => Promise;
+  sortFieldsDict?: { [key: string]: SortFields };
+  sortFields?: SortFields;
 }