+++ aim/web/ui/src/components/MediaPanel/MediaPanel.tsx
@@ -37,6 +37,8 @@ function MediaPanel({
   actionPanel,
   actionPanelSize,
   tooltipType,
+  sortFieldsDict,
+  sortFields,
 }: IMediaPanelProps): React.FunctionComponentElement<React.ReactNode> {
   const [activePointRect, setActivePointRect] = React.useState<{
     top: number;
@@ -158,7 +160,13 @@ function MediaPanel({
   const mediaSetKey = React.useMemo(
     () => Date.now(),
     // eslint-disable-next-line react-hooks/exhaustive-deps
-    [data, wrapperOffsetHeight, wrapperOffsetWidth, additionalProperties],
+    [
+      data,
+      wrapperOffsetHeight,
+      wrapperOffsetWidth,
+      additionalProperties,
+      sortFieldsDict,
+    ],
   );
 
   React.useEffect(() => {
@@ -215,6 +223,7 @@ function MediaPanel({
                     onListScroll={onListScroll}
                     addUriToList={addUriToList}
                     mediaSetKey={mediaSetKey}
+                    sortFieldsDict={sortFieldsDict}
                     wrapperOffsetHeight={wrapperOffsetHeight - 48}
                     wrapperOffsetWidth={wrapperOffsetWidth}
                     focusedState={focusedState}
@@ -224,6 +233,7 @@ function MediaPanel({
                     tableHeight={tableHeight}
                     tooltip={tooltip}
                     mediaType={mediaType}
+                    sortFields={sortFields}
                   />
                 </div>
                 {tooltipType && (