+++ aim/web/ui/src/pages/ImagesExplore/components/Controls/Controls.tsx
@@ -11,6 +11,8 @@ import ImagePropertiesPopover from 'components/ImagePropertiesPopover';
 
 import { CONTROLS_DEFAULT_CONFIG } from 'config/controls/controlsDefaultConfig';
 
+import SortPopover from 'pages/Metrics/components/Table/SortPopover/SortPopover';
+
 import { IControlProps } from 'types/pages/imagesExplore/components/Controls/Controls';
 
 import './Controls.scss';
@@ -124,6 +126,33 @@ function Controls(
           />
         </div>
       </Tooltip>
+      <div>
+        <ControlPopover
+          title='Images Sorting'
+          anchor={({ onAnchorClick, opened }) => (
+            <Tooltip title='Images Sorting'>
+              <div
+                onClick={onAnchorClick}
+                className={`Controls__anchor ${opened ? 'active' : ''}`}
+              >
+                <Icon
+                  className={`Controls__icon ${opened ? 'active' : ''}`}
+                  name='sort-outside'
+                />
+              </div>
+            </Tooltip>
+          )}
+          component={
+            <SortPopover
+              sortOptions={props.selectOptions}
+              sortFields={props.sortFields}
+              onSort={props.onImagesSortChange}
+              readOnlyFieldsLabel={'GROUP BY'}
+              onReset={props.onImagesSortReset}
+            />
+          }
+        />
+      </div>
     </div>
   );
 }