+++ aim/web/ui/src/services/models/explorer/createAppModel.ts
@@ -53,7 +53,6 @@ import {
   IOnGroupingSelectChangeParams,
   IPanelTooltip,
   ITooltipData,
-  SortField,
 } from 'types/services/models/metrics/metricsAppModel';
 import {
   IMetricTrace,
@@ -164,6 +163,7 @@ import { isSystemMetric } from 'utils/isSystemMetric';
 import setDefaultAppConfigData from 'utils/app/setDefaultAppConfigData';
 import getAppConfigData from 'utils/app/getAppConfigData';
 import { getValue } from 'utils/helper';
+import { SortField } from 'utils/getSortedFields';
 import onChangeTrendlineOptions from 'utils/app/onChangeTrendlineOptions';
 
 import { AppDataTypeEnum, AppNameEnum } from './index';
@@ -895,10 +895,10 @@ function createAppModel(appConfig: IAppInitialConfig) {
           configData?.table?.sortFields?.map(
             (f: SortField) =>
               function (metric: IMetric) {
-                return getValue(metric, f[0], '');
+                return getValue(metric, f.value, '');
               },
           ) ?? [],
-          configData?.table?.sortFields?.map((f: SortField) => f[1]) ?? [],
+          configData?.table?.sortFields?.map((f: SortField) => f.order) ?? [],
         ),
       );
       const uniqParams = _.uniq(params);
@@ -945,6 +945,7 @@ function createAppModel(appConfig: IAppInitialConfig) {
 
       const tableColumns = getMetricsTableColumns(
         params,
+        groupingSelectOptions,
         data[0]?.config,
         configData.table?.columnsOrder!,
         configData.table?.hiddenColumns!,
@@ -1022,6 +1023,7 @@ function createAppModel(appConfig: IAppInitialConfig) {
 
       const tableColumns = getMetricsTableColumns(
         params,
+        groupingSelectOptions,
         data[0]?.config,
         configData.table?.columnsOrder!,
         configData.table?.hiddenColumns!,
@@ -1460,6 +1462,7 @@ function createAppModel(appConfig: IAppInitialConfig) {
       );
       const tableColumns: ITableColumn[] = getMetricsTableColumns(
         params,
+        groupingSelectOptions,
         data[0]?.config,
         config?.table?.columnsOrder!,
         config?.table?.hiddenColumns!,
@@ -1624,11 +1627,23 @@ function createAppModel(appConfig: IAppInitialConfig) {
       onResetConfigData({ model, getConfig, updateModelData });
     }
 
-    function onSortChange(
-      field: string,
-      value?: 'asc' | 'desc' | 'none',
-    ): void {
-      onTableSortChange({ field, model, appName, updateModelData, value });
+    function onSortChange({
+      sortFields,
+      order,
+      index,
+      actionType,
+      field,
+    }: any): void {
+      onTableSortChange({
+        field,
+        sortFields,
+        order,
+        index,
+        actionType,
+        model,
+        appName,
+        updateModelData,
+      });
     }
 
     function setModelComponentRefs(refElement: object): void {
@@ -3451,10 +3466,10 @@ function createAppModel(appConfig: IAppInitialConfig) {
             configData?.table?.sortFields?.map(
               (f: SortField) =>
                 function (run: IParam) {
-                  return getValue(run, f[0], '');
+                  return getValue(run, f.value, '');
                 },
             ) ?? [],
-            configData?.table?.sortFields?.map((f: SortField) => f[1]) ?? [],
+            configData?.table?.sortFields?.map((f: SortField) => f.order) ?? [],
           ),
         );
         const uniqParams = _.uniq(params);
@@ -3693,11 +3708,23 @@ function createAppModel(appConfig: IAppInitialConfig) {
         onResetConfigData({ model, getConfig, updateModelData });
       }
 
-      function onSortChange(
-        field: string,
-        value?: 'asc' | 'desc' | 'none',
-      ): void {
-        onTableSortChange({ field, model, appName, updateModelData, value });
+      function onSortChange({
+        sortFields,
+        order,
+        index,
+        actionType,
+        field,
+      }: any): void {
+        onTableSortChange({
+          sortFields,
+          order,
+          index,
+          field,
+          actionType,
+          model,
+          appName,
+          updateModelData,
+        });
       }
 
       function changeLiveUpdateConfig(config: {
@@ -4399,10 +4426,10 @@ function createAppModel(appConfig: IAppInitialConfig) {
             configData?.table?.sortFields?.map(
               (f: SortField) =>
                 function (run: IParam) {
-                  return getValue(run, f[0], '');
+                  return getValue(run, f.value, '');
                 },
             ) ?? [],
-            configData?.table?.sortFields?.map((f: SortField) => f[1]) ?? [],
+            configData?.table?.sortFields?.map((f: SortField) => f.order) ?? [],
           ),
         );
         const uniqParams = _.uniq(params);
@@ -4833,11 +4860,23 @@ function createAppModel(appConfig: IAppInitialConfig) {
         });
       }
 
-      function onSortChange(
-        field: string,
-        value?: 'asc' | 'desc' | 'none',
-      ): void {
-        onTableSortChange({ field, model, appName, updateModelData, value });
+      function onSortChange({
+        sortFields,
+        order,
+        index,
+        actionType,
+        field,
+      }: any): void {
+        onTableSortChange({
+          sortFields,
+          order,
+          index,
+          field,
+          actionType,
+          model,
+          appName,
+          updateModelData,
+        });
       }
 
       function onModelBookmarkCreate({